+++ CHANGELOG.md
@@ -1,4 +1,18 @@
 # CHANGELOG
 
+## 0.3.2
+
+* Add more documentation
+* impl FromStr for ParsedInput as more idiomatic
+* Forgot hcl in one error message
+
+## 0.3.1
+
+* Strip binary
+
+## 0.3.0
+
+* Add HCL support
+
 ## 0.2.0
 * Add reading from stdin